Move isSpatial
