+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/expression/EsqlTypeResolutions.java
@@ -12,7 +12,6 @@ import org.elasticsearch.xpack.esql.core.expression.FieldAttribute;
 import org.elasticsearch.xpack.esql.core.expression.TypeResolutions;
 import org.elasticsearch.xpack.esql.core.type.DataType;
 import org.elasticsearch.xpack.esql.core.type.EsField;
-import org.elasticsearch.xpack.esql.type.EsqlDataTypes;
 
 import java.util.Locale;
 
@@ -67,7 +66,7 @@ public class EsqlTypeResolutions {
     private static final String[] NON_SPATIAL_TYPE_NAMES = DataType.types()
         .stream()
         .filter(DataType::isRepresentable)
-        .filter(t -> EsqlDataTypes.isSpatial(t) == false)
+        .filter(t -> DataType.isSpatial(t) == false)
         .map(DataType::esType)
         .toArray(String[]::new);
 
@@ -76,11 +75,11 @@ public class EsqlTypeResolutions {
     }
 
     public static Expression.TypeResolution isSpatial(Expression e, String operationName, TypeResolutions.ParamOrdinal paramOrd) {
-        return isType(e, EsqlDataTypes::isSpatial, operationName, paramOrd, SPATIAL_TYPE_NAMES);
+        return isType(e, DataType::isSpatial, operationName, paramOrd, SPATIAL_TYPE_NAMES);
     }
 
     public static Expression.TypeResolution isNotSpatial(Expression e, String operationName, TypeResolutions.ParamOrdinal paramOrd) {
-        return isType(e, t -> EsqlDataTypes.isSpatial(t) == false, operationName, paramOrd, NON_SPATIAL_TYPE_NAMES);
+        return isType(e, t -> DataType.isSpatial(t) == false, operationName, paramOrd, NON_SPATIAL_TYPE_NAMES);
     }
 
 }
